ai/services/classification/pyproject.toml
[project]
name = "civitro-classification"
version = "0.1.0"
description = "AI Classification / CivitroVision service"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "Pillow>=10.0.0",
    "numpy>=1.26.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
